+++ Controls/_grid/GridView.ts
@@ -30,7 +30,7 @@ import * as ColumnTpl from 'wml!Controls/_grid/layout/common/ColumnContent';
 import * as GroupTemplate from 'wml!Controls/_grid/GroupTemplate';
 
 import {Logger} from 'UI/Utils';
-import { stickyLadderCellsCount } from 'Controls/_grid/utils/GridLadderUtil';
+import { GridLadderUtil } from 'Controls/display';
 import {SyntheticEvent} from "Vdom/Vdom";
 
 var
@@ -59,7 +59,7 @@ var
             }
             let initialWidths = columns.map(((column) => column.width || GridLayoutUtil.getDefaultColumnWidth()));
             let columnsWidths: string[] = [];
-            const stickyCellsCount = stickyLadderCellsCount(columns, self._options.stickyColumn, self._options.listModel.getDragItemData());
+            const stickyCellsCount = GridLadderUtil.stickyLadderCellsCount(columns, self._options.stickyColumn, self._options.listModel.getDragItemData());
             if (stickyCellsCount === 1) {
                 columnsWidths = ['0px'].concat(initialWidths);
             } else if (stickyCellsCount === 2) {
