+++ Controls/display.ts
@@ -78,6 +78,10 @@ export {ColumnsCollection};
 import {default as ColumnsCollectionItem} from './_display/ColumnsCollectionItem';
 export {ColumnsCollectionItem};
 
+import * as GridLadderUtil from './_display/utils/GridLadderUtil';
+export {GridLadderUtil};
+import isFullGridSupport from './_display/utils/GridSupportUtil';
+export {isFullGridSupport};
 import {default as GridCollection} from './_display/GridCollection';
 export {GridCollection};
 import {default as GridCollectionItem} from './_display/GridCollectionItem';
