+++ Controls/gridNew.ts
@@ -1,5 +1,3 @@
-import { GridLadderUtil } from 'Controls/grid';
-
 import { default as View } from 'Controls/_gridNew/Grid';
 
 import * as ItemTemplate from 'wml!Controls/_gridNew/Render/grid/Item';
@@ -18,6 +16,5 @@ export {
     ColumnTemplate,
     StickyLadderColumnTemplate,
     HeaderContent,
-    FooterContent,
-    GridLadderUtil
+    FooterContent
 };
